# budgetly_api/src/main/resources/api/budgetly-openapi.yml
openapi: 3.0.3
info:
  title: Budgetly API
  description: >
    Family Harmony Budget App — Collaborative household expense tracking
    with AI-powered SMS parsing, category budgets, and flexible N-member splits.
  version: 1.0.0
  contact:
    name: Budgetly Team
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
...
paths:
...
  /auth/google:
    post:
      tags: [Auth]
      operationId: googleSignIn
      summary: Exchange Google ID token for app JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoogleSignInRequest"
      responses:
        "200":
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid Google token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshToken
      summary: Refresh JWT token
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      tags: [Auth]
      operationId: getCurrentUser
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─────────────────────── FAMILY GROUPS ─────────────────────
...
  /messages/process:
    post:
      tags: [Messages]
      operationId: processMessage
      summary: Submit unprocessed SMS for LLM analysis
      description: >
        Sends raw SMS text to the server for Gemini LLM analysis.
        If the message is financial, validates & posts the transaction first.
        Only if transaction validation passes, the LLM-generated regex is saved to the pattern registry.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessMessageRequest"
      responses:
        "200":
          description: Message analyzed by LLM
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessMessageResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /messages/pending:
    get:
      tags: [Messages]
      operationId: getPendingMessages
      summary: Get pending messages for triage (incomplete or unprocessed)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Pending messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
  /messages/ignored:
    get:
      tags: [Messages]
      operationId: getIgnoredMessages
      summary: Get non-financial / deleted messages
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Ignored messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
  /messages/{id}/confirm:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    post:
      tags: [Messages]
      operationId: confirmMessage
      summary: Confirm parsed transaction
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Message confirmed, transaction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /messages/{id}/reject:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    post:
      tags: [Messages]
      operationId: rejectMessage
      summary: Reject / ignore message
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Message rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /messages/{id}/restore:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    post:
      tags: [Messages]
      operationId: restoreMessage
      summary: Restore an ignored message
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Message restored to pending
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # ─────────────────── PATTERN REGISTRY ───────────────────────
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─────────────────── PARAMETERS ────────────────────────────
...
  schemas:
...
    GoogleSignInRequest:
      type: object
      required: [idToken]
      properties:
        idToken:
          type: string
          description: Google ID token from client-side sign-in
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          description: Token expiry in seconds
        user:
          $ref: "#/components/schemas/User"

    # ── User ──
    User:
      type: object
      properties:
        id:
          type: string
        googleId:
          type: string
        displayName:
          type: string
        email:
          type: string
          format: email
        avatarUrl:
          type: string
          format: uri
        createdAt:
          type: string
          format: date-time

    # ── Family Group ──
...
    Category:
      type: object
      properties:
        id:
          type: string
        familyGroupId:
          type: string
        name:
          type: string
        icon:
          type: string
        budgetLimit:
          type: number
          format: double
        color:
          type: string
          description: Hex color code
        sortOrder:
          type: integer
...
    Transaction:
      type: object
      properties:
        id:
          type: string
        familyGroupId:
          type: string
        createdByUserId:
          type: string
        createdByUser:
          $ref: "#/components/schemas/User"
        categoryId:
          type: string
        category:
          $ref: "#/components/schemas/Category"
        amount:
          type: number
          format: double
        currency:
          type: string
          default: INR
        merchant:
          type: string
        description:
          type: string
        type:
          $ref: "#/components/schemas/TransactionType"
        sourceType:
          $ref: "#/components/schemas/SourceType"
        sourceRawText:
          type: string
        matchedPatternId:
          type: string
          description: ID of the regex pattern that parsed this transaction (if sourceType is REGEX_LOCAL)
        notes:
          type: string
        transactionDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    TransactionType:
      type: string
      enum: [EXPENSE, INCOME, SUBSCRIPTION]
    SplitMethod:
      type: string
      enum: [PERSONAL, PERCENT, SHARE]
    SourceType:
      type: string
      enum: [MANUAL, REGEX_LOCAL, LLM_SERVER]
    CreateTransactionRequest:
      type: object
      required: [amount, merchant, type, transactionDate]
      properties:
        amount:
          type: number
          format: double
        currency:
          type: string
          default: INR
        merchant:
          type: string
        description:
          type: string
        type:
          $ref: "#/components/schemas/TransactionType"
        categoryId:
          type: string
        splitMethod:
          $ref: "#/components/schemas/SplitMethod"
        splitConfig:
          type: object
          additionalProperties:
            type: number
        notes:
          type: string
        transactionDate:
          type: string
          format: date-time
...
    Message:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        familyGroupId:
          type: string
        sender:
          type: string
        rawText:
          type: string
        status:
          $ref: "#/components/schemas/MessageStatus"
        parseSource:
          $ref: "#/components/schemas/ParseSource"
        matchedPatternId:
          type: string
          description: ID of the regex pattern that matched (if parseSource is REGEX_LOCAL)
        parsedTransaction:
          $ref: "#/components/schemas/CreateTransactionRequest"
        linkedTransactionId:
          type: string
        receivedAt:
          type: string
          format: date-time
    MessageStatus:
      type: string
      enum: [PENDING, CONFIRMED, REJECTED, IGNORED]
    ParseSource:
      type: string
      enum: [REGEX_LOCAL, LLM_SERVER, MANUAL]
    ProcessMessageRequest:
      type: object
      required: [sender, rawText, familyGroupId]
      properties:
        sender:
          type: string
          description: SMS sender ID (e.g. "HDFCBK", "SBIUPI")
        rawText:
          type: string
          description: Full raw SMS text
        familyGroupId:
          type: string
    ProcessMessageResponse:
      type: object
      properties:
        isFinancial:
          type: boolean
          description: Whether the LLM classified this message as financial
        parsedData:
          $ref: "#/components/schemas/CreateTransactionRequest"
        generatedPattern:
          $ref: "#/components/schemas/SmsPattern"
        transactionId:
          type: string
          description: ID of the created transaction (if financial and validation passed)
        message:
          $ref: "#/components/schemas/Message"

    # ── SMS Pattern Registry ──
    SmsPattern:
      type: object
      properties:
        id:
          type: string
        sender:
          type: string
          description: SMS sender ID (e.g. "HDFCBK", "SBIUPI", "ICICIB")
        regex:
          type: string
          description: Java/Dart-compatible regex pattern with named capture groups
        extractionMap:
          $ref: "#/components/schemas/ExtractionMap"
        sampleMessage:
          type: string
          description: The original SMS that generated this pattern
        usageCount:
          type: integer
          format: int64
          description: Global counter of successful matches across all users
        createdAt:
          type: string
          format: date-time
    ExtractionMap:
      type: object
      description: Maps regex capture groups to transaction fields. amount, merchant, and timestamp are mandatory.
      required: [amount, merchant, timestamp]
      properties:
        amount:
          type: string
          description: Capture group name for amount
        merchant:
          type: string
          description: Capture group name for merchant
        timestamp:
          type: string
          description: Capture group name for timestamp
        accountLast4:
          type: string
          description: Capture group name for last 4 digits of account (optional)
        type:
          type: string
          description: Capture group name for debit/credit type (optional)
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string